Refactor async PPO: split into main/train/utils, mujoco-torch only
- ppo_async_mujoco.py: stripped to just main() with top-level imports
- train.py: new file with train_start() and train_iterate() loops
- utils_mujoco.py: rewritten for mujoco-torch only (no gymnasium),
  VecNormV2, 4x256 MLPs, all utility classes (ActorWithCritic,
  WorkerGAEPostproc, LearnerPostproc, RenderFirstEnv) moved here
- config_mujoco.yaml: removed backend/num_workers fields
